+++ CHANGELOG.md
@@ -7,12 +7,12 @@
 - Default stackableVersion to operator version ([#385]).
 - Configuration overrides for the JVM security properties, such as DNS caching ([#389]).
 - Support PodDisruptionBudgets ([#399]).
+- Support graceful shutdown ([#402]).
 - Added support for version 2.4.17 ([#403]).
 
 ### Changed
 
-- `vector` `0.26.0` -> `0.33.0` ([#382]).
-- `operator-rs` `0.44.0` -> `0.55.0` ([#385], [#399], [#403]).
+- `vector` `0.26.0` -> `0.31.0` ([#382]).
 - Use jmx_exporter soft link instead of hardcoded version ([#403]).
 
 ### Fixed
@@ -28,6 +28,7 @@
 [#389]: https://github.com/stackabletech/hbase-operator/pull/389
 [#394]: https://github.com/stackabletech/hbase-operator/pull/394
 [#399]: https://github.com/stackabletech/hbase-operator/pull/399
+[#402]: https://github.com/stackabletech/hbase-operator/pull/402
 [#403]: https://github.com/stackabletech/hbase-operator/pull/403
 
 ## [23.7.0] - 2023-07-14